+++ packages/core/src/parsers/jsx/element-parser.ts
@@ -290,7 +290,9 @@ export const jsxElementToJson = (
 
           memo.bindings[key] = createSingleBinding({
             code: generate(expression.body, { compact: true }).code,
+            async: expression.async === true ? true : undefined,
             arguments: args.length ? args : undefined,
+            bindingType: 'function',
           });
         } else if (types.isJSXElement(expression)) {
           // <Foo myProp={<MoreMitosisNode><div /></MoreMitosisNode>} />