+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @builder.io/mitosis-cli
 
+## 0.5.15
+
+### Patch Changes
+
+- Updated dependencies [a0ad5ab]
+  - @builder.io/mitosis@0.5.15
+
+## 0.5.14
+
+### Patch Changes
+
+- Updated dependencies [39af4d6]
+  - @builder.io/mitosis@0.5.14
+
+## 0.5.13
+
+### Patch Changes
+
+- Updated dependencies [c7d2f8c]
+  - @builder.io/mitosis@0.5.13
+
 ## 0.5.12
 
 ### Patch Changes